+++ dash_labs/__init__.py
@@ -2,3 +2,5 @@ from . import plugins
 from .version import __version__
 from .dashdown import MarkdownAIO
 from dash_labs import css
+
+from . util import print_registry